+++ Fox/Public/Reporters/FOXDebugReporter.m
@@ -49,13 +49,13 @@
 
 - (void)runnerDidPassNumberOfTests:(NSUInteger)testNumber withResult:(FOXRunnerResult *)result
 {
-    [self logString:[NSString stringWithFormat:@"\n\n%lu Tests Passed.", testNumber]];
+    [self logString:[NSString stringWithFormat:@"\n\n%lu Tests Passed.", (unsigned long)testNumber]];
 }
 
 - (void)runnerDidFailTestNumber:(NSUInteger)testNumber withResult:(FOXRunnerResult *)result
 {
     [self logString:[NSString stringWithFormat:@"\n\n  %@\n", [[result friendlyDescription] stringByReplacingOccurrencesOfString:@"\n" withString:@"\n  "]]];
-    [self logString:[NSString stringWithFormat:@")\n\nFailure after %lu tests.\n", testNumber + 1]];
+    [self logString:[NSString stringWithFormat:@")\n\nFailure after %lu tests.\n", (unsigned long)(testNumber + 1)]];
 }
 
 - (void)runnerDidRunWithResult:(FOXRunnerResult *)result
